Reorganise document submission fixtures
