+++ cypress/fixtures/document_submissions/get.json
@@ -1,0 +1,18 @@
+{
+  "documentType": {
+    "id": "passport-scan",
+    "title": "Passport",
+    "description": "A valid passport open at the photo page"
+  },
+  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
+  "createdAt": "2021-01-14T10:23:42.958Z",
+  "claimId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
+  "rejectionReason": "string",
+  "state": "PENDING",
+  "document": {
+    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
+    "createdAt": "2021-01-29T16:28:11.326Z",
+    "fileSize": 25300,
+    "fileType": "image/png"
+  }
+}
