+++ src/boundary/response-mapper.test.ts
@@ -1,7 +1,7 @@
 import { DateTime } from 'luxon';
 import EvidenceRequestFixture from '../../cypress/fixtures/evidence_requests/index.json';
 import DocumentTypeFixture from '../../cypress/fixtures/document_types/index.json';
-import DocumentSubmissionFixture from '../../cypress/fixtures/document_submissions/id.json';
+import DocumentSubmissionCreate from '../../cypress/fixtures/document_submissions/create.json';
 import { DeliveryMethod, EvidenceRequest } from '../domain/evidence-request';
 import { Resident } from '../domain/resident';
 import { ResponseMapper } from './response-mapper';
